# scx_lavd: Hardening to avoid stalls when extremely over-loaded.