Harden the deadline calculation and slice boost when the system is extremely overloaded to avoid stalls. It includes the following tuning and changes that won't affect the normal use cases:

1) Give 2x weight boost for a kernel or pinned task. This helps a pinned kernel
   thread be scheduled before the stall.

2) Reduce LAVD_TARGETED_LATENCY_NS from 20 msec to 10 msec. Under extremely
   heavy load, the scheduler shortens the time slices more aggressively. In the
   normal use up to around 3x overload (one running + 2 waiting tasks with
   5 msec time slice), this should not affect.

3) Reduce LAVD_SLICE_BOOST_MAX from 1 sec to 0.5 sec, which is a more realistic
   maximum value.

4) Previously, the slice boost is triggered when slice_max_ns <=
   sys_stat.slice. This is not ideal since it creates a dependency on the time
   slice calculation. Let's change this sys_stat.nr_queued_task <=
   sys_stat.nr_active. The changed condition is still aggressive enough since
   it will try slice boosting until 2x overload (1 for running + 1 for
   waiting).

The changes tested on a 16-CPU AMD Ryzen 9 PRO 6950H with PREEMPT_DYNAMIC and PREEMPT_RT kernels. The scheduler runs well without stalling while running the following all at the same time:

- Linux kenrel compilation with make -j64
- schbench -L -m 4 -M auto -t 64 -n 0
- perf bench sched messaging -t -g 16 -l 3000000
- stress-ng -c 2048
- nice -n 19 stress-ng -c 2048
- Running WebGL Aquarium on Chromium with 30000 fishes (https://webglsamples.org/aquarium/aquarium.html)

This might fix the issue [1], but it has not been tested due to the lack of hardware.

[1] https://github.com/sched-ext/scx/issues/2548